+++ toshi_hazard_store/v2/db_adapter/test/test_adapter_batched.py
@@ -1,8 +1,14 @@
 # from moto import mock_dynamodb
 # from nzshm_common.location.code_location import CodedLocation
+import os
+
 import pytest
 from moto import mock_dynamodb
 from pytest_lazyfixture import lazy_fixture
+
+
+def test_env(tmp_path):
+    assert os.environ["THS_SQLITE_FOLDER"] == str(tmp_path)
 
 
 @pytest.mark.parametrize(